+++ tests/testthat/test-link-elements.R
@@ -1,7 +1,11 @@
 test_that("link_gh_issue() + markup_href() work", {
   expect_snapshot({
-    lines <- c("Go to rstudio/rstudio#120 and [gh](https://github.com) and {.href [gh](https://github.com)}.")
+    lines <- c(
+      "Go to rstudio/rstudio#120 and [gh](https://github.com) and {.href [gh](https://github.com)}.",
+      "For faster printing of message (r-lib/cli#607), use withr::local_options(list(cli.num_colors = cli::num_ansi_colors()))",
+      "[franchise](https://en.wikipedia.org/wiki/Fallout_(franchise))"
+    )
     lines_with_md_link <- link_gh_issue(lines)
-    cat(markup_href(lines_with_md_link))
+    cat(markup_href(lines_with_md_link), sep = "\n")
   })
 })